[advanced] MonthView 컴포넌트 분리
